move hexes to their own module
